Use CompletableFuture to configure device after successful command sent
